completed the controller for holding

+++ app/src/main/java/io/eternalwind/mira/app/dtos/HoldingDto.java
@@ -1,6 +1,5 @@
 package io.eternalwind.mira.app.dtos;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 import lombok.experimental.UtilityClass;
@@ -9,15 +8,15 @@ import lombok.experimental.UtilityClass;
 public class HoldingDto {
     public static record Create(String ticker,
                                 String name,
-                                BigDecimal amount,
+                                Double amount,
                                 String sector) {
     }
 
     public static record Read(UUID id, 
                               String ticker, 
                               String name, 
-                              BigDecimal amount, 
-                              BigDecimal price, 
+                              Double amount, 
+                              Double price, 
                               String sector, 
                               String createdTime,
                               String lastUpdatedTime) {
@@ -25,8 +24,8 @@ public class HoldingDto {
 
     public static record Update(String ticker, 
                                 String name, 
-                                BigDecimal amount, 
-                                BigDecimal price,
+                                Double amount, 
+                                Double price,
                                 String sector) {
     }
 }
